Code Fixed

+++ battleship/battleships/src/Model/Ship.cs
@@ -1,9 +1,4 @@
-
-using Microsoft.VisualBasic;
-using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Diagnostics;
 
 /// <summary>
 /// A Ship has all the details about itself. For example the shipname,
@@ -96,7 +91,7 @@ public class Ship
 		_tiles = new List<Tile>();
 
 		//gets the ship size from the enumarator
-		_sizeOfShip = _shipName;
+		_sizeOfShip =(int) _shipName;
 	}
 
 	/// <summary>
